backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
aiohttp==3.9.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
celery==5.3.4
python-multipart==0.0.6
httpx==0.26.0
python-telegram-bot>=20.0
requests>=2.31.0
ccxt>=4.0.0
pandas>=1.5.0
numpy>=1.24.0
TA-Lib>=0.4.28
websockets>=11.0.0
asyncio>=3.4.3
scikit-learn>=1.3.0  # Для ML моделей